# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Shop POS API Documentation",
        "description": "Swagger OpenApi description for the Coffee Shop POS system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Dynamic API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/products": {
            "get": {
                "tags": [
                    "Products Management"
                ],
                "summary": "List Products",
                "description": "Retrieve a paginated list of products. Requires `admin.catalog` permission.",
                "operationId": "23684014e1fa3cbecd81f67af692ace3",
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Caffe Latte"
                                                            },
                                                            "category_id": {
                                                                "type": "string",
                                                                "example": "uuid"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 4.5
                                                            },
                                                            "sku": {
                                                                "type": "string",
                                                                "example": "HOT-LAT-01"
                                                            },
                                                            "is_active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "id": {
                                                                "type": "string",
                                                                "example": "uuid"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2023-10-01T12:00:00Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "example": "2023-10-01T12:00:00Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products Management"
                ],
                "summary": "Create Product",
                "description": "Create a new Product. Requires `admin.catalog` permission.",
                "operationId": "06cfb0e0c461d62127f9f64d986de44d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "branch_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Caffe Latte"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Delicious latte"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "HOT-LAT-01"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/products/{id}": {
            "get": {
                "tags": [
                    "Products Management"
                ],
                "summary": "Get Product",
                "description": "Retrieve a single Product by ID. Requires `admin.catalog` permission.",
                "operationId": "0f14d80ef7c6bb43727a64aa41782cc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Details retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Caffe Latte"
                                                },
                                                "category_id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 4.5
                                                },
                                                "sku": {
                                                    "type": "string",
                                                    "example": "HOT-LAT-01"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-10-01T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-10-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products Management"
                ],
                "summary": "Update Product",
                "description": "Update an existing Product. Requires `admin.catalog` permission.",
                "operationId": "737c687dac40d21c9ce7a516f1932171",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Caffe Latte"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Delicious latte"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "HOT-LAT-01"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products Management"
                ],
                "summary": "Delete Product",
                "description": "Soft-delete a Product. Requires `admin.catalog` permission.",
                "operationId": "7d4ebe2a1e082a9bfb5ac3beee7579d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}